forrent/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
python-dotenv==1.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
channels==4.0.0
django-allauth==0.52.0